refactor: change frame rate from f64 to u64 and add instruction throttling
- Convert frame_rate type from f64 to u64 across codebase
- Add instruction throttling with 700 instructions per second
- Execute multiple instructions per frame based on frame rate
- Reduce event polling timeout from 16ms to 5ms for better responsiveness
- Maintain consistent typing for frame rate calculations

+++ src/main.rs
@@ -11,7 +11,7 @@ mod state;
 #[command(version, about, long_about = None)]
 struct Args {
     #[arg(short, long, default_value_t = DEFAULT_FRAME_RATE, help = "Frame rate in frames per second")]
-    frame_rate: f64,
+    frame_rate: u64,
 
     #[arg(short, long, help = "Path to the ROM file to run")]
     rom_path: String,
